Add the equivalent functionality of erase_redirect_results
Without this after we respond with a redirect the actual response
status is a 302 instead of a 200
(We want to inject the redirection in a script and let the browser run it).
For some reason after the plugin removed the call to
erase_redirect_results (no longer supported in rails 3) it did not add
the equivalent functionality.

See:
http://apidock.com/rails/ActionController/Base/erase_redirect_results